Prevent from crashes and possible undefined behaviour.
The V8 Engine requires the arguments passed to it for storing not to be
null. However, current code allows those to be set and the code using
the v8-rs may (and does) set the NULL pointers where it is prohibited.

This commit adds checks for those necessary arguments not being null
where those shouldn't.