Introduce strict index types for V8 storage and improve DataGuard.
This ensures compile-time safety and is zero-cost as the compiler
optimises everything out, what leads to just having integers in the end,
but always the proper and intended ones.

Also improves the *DataGuards so that they keep track of the object
which is stored in V8. Note that for the types implementing Copy there
will be no need in keeping track of the life time of the objects as it
makes no sense and compiler won't do that.