Fix leak on embeder data
## Leak reproduction:

Load the function:
```
> redis-cli
127.0.0.1:6379> RG.FUNCTION LOAD UPGRADE "#!js name=lib1\n redis.register_function('foo', ()=>{return redis.v8_version();});"
OK
```

Run redis-benchmark:
```
redis-benchmark -n 10000000 -P 10 RG.FCALL lib foo 0
```

Memory will keep increasing forever.

## Leak root cause:

It is seems that the leak is cause by creating external JS objects (`v8::External`).
`v8::External` allows the embeder to store a `void*` to external data that the embeder might need.
It is seems that exteranl data are stored in some Map which is never cleaned.

The embeder data that was leaking is the indication whether or not the Redis GIL is held.

### Short term solution

Reuse the same `v8::External` instead of creating a new one each time. The `v8::External` will hold a pointer
to an wrapper structur that will hold a pointer to the actual embeder data. If we want to reset this pointer
we simply set the wrapper to point to NULL but we do not unset the `v8::External` object.

### Short term solution

We need to investigate further why this leak is happening and determind if there are more places that might be
effected by it (we use `v8::External` in more places). A seperate issue to continue the investigation will be opened.