Support optional parameter on NativeFunctionArgument derive macro. (#25)
The PR extends `NativeFunctionArgument` derive macro to also support and optional argument. So it will be possible to write the following native function and have the 2 last arguments as optional:

```rust
   #[derive(NativeFunctionArgument, PartialEq, Eq, Debug)]
    struct InnerArgs {
        i: i64,
    }

    #[derive(NativeFunctionArgument, PartialEq, Eq, Debug)]
    struct Args {
        i: i64,
        s: String,
        b: bool,
        o: Option<String>,
        inner: InnerArgs,
        optional_inner: Option<InnerArgs>,
    }

   new_native_function!(|_isolate, _ctx_scope, args: Args, optional_args1: Option<Args>,optional_args2: Option<Args>| {}
```

The following call will match to the above native function:

```js
test({i: 1, s: 'foo', b: false, inner: { i: 10 }}, {i: 1, s: 'foo', b: false, inner: { i: 10 }})
```

This is useful when we want to have a final argument as a optional JS object.