Fix use after free of native function return value.
When we call a native function we should avoid creating a new
isolate scope because when the isolate scope will be freed, all
the local handlers will be marked for GC and potenatially freed
as well. This includes the return value.

To avoid that, we create a dummy isolate scope. From user POV,
the dummy isolate scope can be used like a regular isolate scope,
But it will not cause local handler to be freed when it is released.