Added on_dropped callback to V8LocalValue. (#40)
Added an option to set an on_dropped callback on V8LocalValue. The given callback will be called when the GC will decide to collect the value. To test the new functionality, added an API to run GC for testing. We use this new API to force the GC to free the V8LocalValue and call our callback.