Seperate patform initialisation from V8 initialisation. (#45)
Seperate patform initialisation from V8 initialisation.

The platform initialisation must happened on the process main thread,
otherwise it causes crashes if later, trying to run JS code on the main thread.

Initialising platform is cheep and can happened on processes start. Initialise the V8
can be heavy and might requere to happened only when V8 is in fact in use.

The PR Splits the platform initialisation from the V8 initialisation so the user can initialise
them separately.