Add InspectorGuard with some critical inspector interface.
This moves some of the functionality of the Inspector struct to the new
InspectorGuard struct, objects of which may only be created when the
isolate scope passed is for the same isolate this Inspector object was
created with. This is done to make sure that the APIs which rely on the
isolate and the corresponding HandleScope are correctly used.