Add cargo deny configuration (#64)
* Correctly specify the license.

* Add cargo-deny toml configuration.

Adds the cargo-deny configuration file for the project. It allows to
check for license compatibilities of the project and its dependencies,
as well as other checks, one of the most important ones of which is the
check for the binaries.

* Add CI checks for cargo-deny.

Adds the use of cargo-deny to automatically catch any issues before even
building the crate(s).